Migrated database to PostgreSQL from MySQL.

+++ src/controllers/logs.controller.ts
@@ -11,6 +11,7 @@ export default class LogsController {
   constructor(dbConn: any) {
     this.service = new LogService(dbConn);
 
+    //TODO: Change this to post and get session ID from cookie.
     this.routes.get("/", async (req: Request, res: Response) => {
       const id: any = req.query.id;
       const decend: any = req.query.decend;
@@ -18,21 +19,28 @@ export default class LogsController {
       const userId = await this.service
         .fetchUserId(id)
         .catch((e) => console.log(e));
-
-      res.send(await this.service.getLog(userId, decend));
+      if (userId) {
+        res.send(await this.service.getLog(userId, decend));
+      } else {
+        res.send("Fetching Log Failed");
+      }
     });
 
     this.routes.get("/getrecord", async (req: Request, res: Response) => {
-      const id: any = req.query.uid;
+      const id: any = req.query.uid; //this is actually the session id
       const rid: any = req.query.rid;
 
       const userId = await this.service
         .fetchUserId(id)
         .catch((e) => console.log(e));
 
-      const record: Record = await this.service.getRecord(userId, rid);
+      if (userId) {
+        const record: Record = await this.service.getRecord(userId, rid);
 
-      res.send(record);
+        res.send(record);
+      } else {
+        res.send("Fetching Record Failed");
+      }
     });
 
     this.routes.post("/addrecord", async (req: Request, res: Response) => {
@@ -42,11 +50,15 @@ export default class LogsController {
         .fetchUserId(newRecord.userId)
         .catch((e) => console.log(e));
 
-      const newRecordConfirm = await this.service
-        .addRecord(newRecord, userId)
-        .catch((e) => console.log(e));
+      if (userId) {
+        const newRecordConfirm = await this.service
+          .addRecord(newRecord, userId)
+          .catch((e) => console.log(e));
 
-      res.send(newRecordConfirm);
+        res.send(newRecordConfirm);
+      } else {
+        res.send("Add Record Failed");
+      }
     });
 
     this.routes.post("/editrecord", async (req: Request, res: Response) => {
@@ -57,11 +69,15 @@ export default class LogsController {
         .fetchUserId(uid)
         .catch((e) => console.log(e));
 
-      const edited = await this.service
-        .editRecord(record, userId)
-        .catch((e) => console.log(console.log(e)));
+      if (userId) {
+        const edited = await this.service
+          .editRecord(record, userId)
+          .catch((e) => console.log(console.log(e)));
 
-      res.send(edited);
+        res.send(edited);
+      } else {
+        res.send("Editing Record Failed");
+      }
     });
 
     this.routes.post("/deleterecord", async (req: Request, res: Response) => {
@@ -72,14 +88,18 @@ export default class LogsController {
         .fetchUserId(uid)
         .catch((e) => console.log(e));
 
-      const deleted = await this.service
-        .deleteRecord(userId, recordId)
-        .catch((e) => console.log(console.log(e)));
+      if (userId) {
+        const deleted = await this.service
+          .deleteRecord(userId, recordId)
+          .catch((e) => console.log(console.log(e)));
 
-      res.send(deleted);
+        res.send(deleted);
+      } else {
+        res.send("Delete Record Failed");
+      }
     });
 
-    this.routes.get("/async", async (req: Request, res: Response) => {
+    /*this.routes.get("/async", async (req: Request, res: Response) => {
       const id: any = req.query.id;
 
       //result from db query is passed as a parameter to the callback function,
@@ -87,6 +107,6 @@ export default class LogsController {
       await this.service.aGetLog(id, (result: any) => {
         res.send(result);
       });
-    });
+    });*/
   }
 }
